move tests into separate files
