+++ packages/react-hig/src/Button.test.js
@@ -1,0 +1,182 @@
+import { mount } from 'enzyme';
+import * as HIG from 'hig.web';
+import React from 'react';
+
+import { Button } from './react-hig';
+
+describe('<Button>', () => {
+  /**
+     * Creates a hig.web button and returns the instance and it's container
+     *
+     * @param {object} defaults hig.web defaults for the button
+     */
+  function createHigButton(defaults = {}) {
+    const higContainer = document.createElement('div');
+
+    // use spread here to clone defaults since HIG.Button mutates this object
+    const higButton = new HIG.Button({ ...defaults });
+
+    higButton.mount(higContainer);
+
+    return { higButton, higContainer };
+  }
+
+  it('renders the standard button', () => {
+    const defaults = { title: 'regular button', link: 'http://example.com' };
+
+    const { higButton, higContainer } = createHigButton(defaults);
+
+    const container = document.createElement('div');
+
+    const wrapper = mount(<Button {...defaults} />, { attachTo: container });
+
+    expect(container.firstChild.outerHTML).toMatchSnapshot();
+
+    expect(container.firstChild.outerHTML).toEqual(
+      higContainer.firstChild.outerHTML
+    );
+  });
+
+  it('updates the title the same as setTitle', () => {
+    const { higButton, higContainer } = createHigButton();
+
+    const newTitle = 'correct title';
+
+    // update via hig API
+    higButton.setTitle(newTitle);
+
+    const container = document.createElement('div');
+
+    // update via React API
+    const wrapper = mount(<Button />, { attachTo: container });
+
+    wrapper.setProps({ title: newTitle });
+
+    expect(container.firstChild.outerHTML).toMatchSnapshot();
+
+    expect(container.firstChild.outerHTML).toEqual(
+      higContainer.firstChild.outerHTML
+    );
+  });
+
+  it('removes the title if it is no longer specified', () => {
+    const defaults = { title: 'some title' };
+
+    const { higButton, higContainer } = createHigButton(defaults);
+
+    const reactContainer = document.createElement('div');
+
+    const wrapper = mount(<Button {...defaults} />, {
+      attachTo: reactContainer
+    });
+
+    higButton.setTitle(null);
+
+    // update to remove the title prop
+    wrapper.setProps({ title: undefined });
+
+    expect(reactContainer.firstChild.outerHTML).toMatchSnapshot();
+
+    expect(reactContainer.firstChild.outerHTML).toEqual(
+      higContainer.firstChild.outerHTML
+    );
+  });
+
+  it('updates the link the same as setLink', () => {
+    const { higButton, higContainer } = createHigButton();
+    const newLink = 'http://example.com/2';
+
+    higButton.setLink(newLink);
+
+    const reactContainer = document.createElement('div');
+
+    const wrapper = mount(<Button />, { attachTo: reactContainer });
+
+    wrapper.setProps({ link: newLink });
+
+    expect(reactContainer.firstChild.outerHTML).toMatchSnapshot();
+
+    expect(reactContainer.firstChild.outerHTML).toEqual(
+      higContainer.firstChild.outerHTML
+    );
+  });
+
+  ['onClick', 'onHover'].forEach(eventName => {
+    it(`sets up ${eventName} initially`, () => {
+      const eventSpy = jest.fn();
+
+      const reactContainer = document.createElement('div');
+
+      const wrapper = mount(<Button {...{ [eventName]: eventSpy }} />, {
+        attachTo: reactContainer
+      });
+
+      // click on the rendered button
+      const instance = wrapper.instance().instance;
+
+      // This is the same fn we pass to hig.web
+      instance.events[eventName]();
+
+      // expect onClickSpy to be called
+      expect(eventSpy).toBeCalled();
+    });
+
+    it(`updates ${eventName} when it updates`, () => {
+      const eventSpy1 = jest.fn();
+      const eventSpy2 = jest.fn();
+
+      const reactContainer = document.createElement('div');
+
+      const wrapper = mount(<Button {...{ [eventName]: eventSpy1 }} />, {
+        attachTo: reactContainer
+      });
+
+      wrapper.setProps({ [eventName]: eventSpy2 });
+
+      // click on the rendered button
+      const instance = wrapper.instance().instance;
+
+      // This is the same fn we pass to hig.web
+      instance.events[eventName]();
+
+      // expect onClickSpy to be called
+      expect(eventSpy2).toBeCalled();
+    });
+
+    it(`updates ${eventName} when it wasn't specified to begin with`, () => {
+      const eventSpy1 = jest.fn();
+
+      const reactContainer = document.createElement('div');
+
+      const wrapper = mount(<Button />, {
+        attachTo: reactContainer
+      });
+
+      wrapper.setProps({ [eventName]: eventSpy1 });
+
+      const instance = wrapper.instance().instance;
+
+      // This is the same fn we pass to hig.web
+      instance.events[eventName]();
+
+      // expect onClickSpy to be called
+      expect(eventSpy1).toBeCalled();
+    });
+
+    it(`removes ${eventName} if it is no longer specified`, () => {
+      const eventSpy1 = jest.fn();
+
+      const reactContainer = document.createElement('div');
+
+      const wrapper = mount(<Button {...{ [eventName]: eventSpy1 }} />, {
+        attachTo: reactContainer
+      });
+
+      wrapper.setProps({ [eventName]: undefined });
+
+      const instance = wrapper.instance().instance;
+
+      expect(instance.events[eventName]).toBeUndefined();
+    });
+  });
+});
